# entity-service/openapi.yaml
openapi: 3.0.1
info:
  title: Entity Service API
  version: 1.0.0
servers:
  - url: "{server}:{port}/"
    variables:
      server:
        default: http://localhost
      port:
        default: "8080"
paths:
...
  /comments:
    post:
      summary: Create a comment on any supported reference entity.
      description: >-
        The reference-generic comment-create path, used for reference types that have
        no dedicated create-comment route of their own: change_request, incident,
        conversation and deployment. referenceType `case` is rejected with 400: case
        comments are owned by POST /cases/{id}/comments, which applies the
        case-specific rules.
      operationId: createComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        "200":
          description: The created comment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCommentResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Referenced entity not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /change-requests:
    post:
      summary: Create a change request (ServiceNow data source only).
      operationId: createChangeRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChangeRequestRequest'
      responses:
        "201":
          description: Change request created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateChangeRequestResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /change-requests/{id}:
    patch:
      summary: Update a change request (ServiceNow data source only).
      operationId: patchChangeRequest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchChangeRequestRequest'
      responses:
        "200":
          description: The updated change request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatchChangeRequestResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Change request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Get a single change request by ID (ServiceNow data source only).
      operationId: getChangeRequest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The change request detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequest'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Change request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /change-requests/{id}/approvals:
    get:
      summary: Get the approval stages for a change request (ServiceNow data source only).
      operationId: getChangeRequestApprovals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The change request's approval stages.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequestApprovals'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Change request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /change-requests/{id}/approvals/decision:
    post:
      summary: Decide the caller's own pending approval on a change request.
      description: >-
        Records the caller's decision on their own pending approval stage. Only the
        caller's own pending approval can be acted on; the change request's own state
        cascades automatically in the backing data source.
      operationId: decideChangeRequestApproval
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeRequestApprovalDecisionRequest'
      responses:
        "200":
          description: The decided approval record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequestApprovalDecisionResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Change request or pending approval not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /change-requests/search:
    post:
      summary: Search change requests (ServiceNow data source only).
      operationId: searchChangeRequests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchChangeRequestsRequest'
      responses:
        "200":
          description: Change requests matching the filters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchChangeRequestsResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /comments/search:
    post:
      summary: Search comments by reference entity (ServiceNow data source only).
      operationId: searchComments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchCommentsRequest'
      responses:
        "200":
          description: Paginated list of comments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchCommentsResponse'
        "400":
          description: Bad request (missing or invalid referenceId/referenceType).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          description: Number of results per page. Defaults to 20, capped at 50.
        offset:
          type: integer
          description: Zero-based offset into the result set.
...
    EntityRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
...
    CreateCommentRequest:
      type: object
      required: [referenceId, referenceType, type, content]
      properties:
        referenceId:
          type: string
          format: uuid
          description: UUID of the entity the comment belongs to.
        referenceType:
          type: string
          enum: [conversation, change_request, deployment, incident]
          description: >-
            Deliberately a strict subset of ReferenceType: `case` is rejected with
            400 here, because case comments are owned by POST /cases/{id}/comments.
            Declared inline rather than referencing the shared ReferenceType, which
            must keep accepting `case` for attachments and comment search.
        type:
          type: string
          enum: [work_note, comment, activity]
          description: work_note is restricted to internal users; activity to internal and system users.
        content:
          type: string
    CreateCommentResponse:
      type: object
      properties:
        message:
          type: string
        comment:
          $ref: '#/components/schemas/CommentDetail'
    CommentDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdOn:
          type: string
          format: date-time
        createdBy:
          type: string
          description: Email of the user who created the comment.
...
    SearchChangeRequestsRequest:
      type: object
      properties:
        filters:
          type: object
          properties:
            projectIds:
              type: array
              items:
                type: string
                format: uuid
            searchQuery:
              type: string
            states:
              type: array
              items:
                type: string
                enum: [customer_approval, scheduled, implement, review, customer_review, rollback, closed, canceled]
            impacts:
              type: array
              items:
                type: string
                enum: [high, medium, low]
            closedStartDate:
              type: string
              format: date-time
              description: "ISO-8601 UTC datetime (YYYY-MM-DDTHH:MM:SSZ)"
            closedEndDate:
              type: string
              format: date-time
              description: "ISO-8601 UTC datetime (YYYY-MM-DDTHH:MM:SSZ)"
        sortBy:
          type: object
          required: [field, order]
          properties:
            field:
              type: string
              enum: [createdOn, updatedOn]
            order:
              type: string
              enum: [asc, desc]
        pagination:
          type: object
          properties:
            offset:
              type: integer
              minimum: 0
              default: 0
            limit:
              type: integer
              minimum: 1
              maximum: 50
              default: 20
...
    SearchChangeRequestsResponse:
      type: object
      properties:
        changeRequests:
          type: array
          items:
            $ref: '#/components/schemas/SearchChangeRequestView'
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
    ChangeRequest:
      allOf:
        - $ref: '#/components/schemas/SearchChangeRequestView'
        - type: object
          properties:
            createdBy:
              type: string
            justification:
              type: string
              nullable: true
            impactDescription:
              type: string
              nullable: true
            serviceOutage:
              type: string
              nullable: true
            communicationPlan:
              type: string
              nullable: true
            rollbackPlan:
              type: string
              nullable: true
            testPlan:
              type: string
              nullable: true
            hasCustomerApproved:
              type: boolean
            hasCustomerReviewed:
              type: boolean
            approvedBy:
              $ref: '#/components/schemas/EntityRef'
              nullable: true
            approvedOn:
              type: string
              nullable: true
...
    ChangeRequestApproval:
      type: object
      properties:
        stage:
          type: string
          description: Approval stage name (e.g. Assess, Authorize, Customer Approval).
        approverType:
          type: string
          enum: [STATIC_GROUP, DYNAMIC_CONTACT]
        approverName:
          type: string
          description: Group name (STATIC_GROUP) or the per-CR customer contact's name (DYNAMIC_CONTACT).
        status:
          type: string
          enum: [APPROVED, REJECTED, PENDING]
        approvers:
          type: array
          items:
            $ref: '#/components/schemas/ChangeRequestApprover'
    ChangeRequestApprovals:
      type: object
      properties:
        approvals:
          type: array
          items:
            $ref: '#/components/schemas/ChangeRequestApproval'
    ChangeRequestApprovalDecisionRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [approved, rejected]
          description: Caller's decision on their own pending approval.
    ChangeRequestApprovalDecisionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: UUID of the approval record that was decided.
        state:
          type: string
          description: Resulting state of the approval ("approved" or "rejected").
...
    CreateChangeRequestRequest:
      type: object
      required:
        - subject
      properties:
        subject:
          type: string
        category:
          type: string
          enum:
            - hardware
            - software
            - service
            - system_software
            - applications_software
            - network
            - telecom
            - documentation
            - other
            - regular_release_cloud
            - hotfix_release_cloud
            - devops
            - cloud_computing
        serviceId:
          type: string
          format: uuid
        serviceOfferingId:
          type: string
          format: uuid
        configurationItemId:
          type: string
          format: uuid
        priority:
          type: string
          enum:
            - critical
            - high
            - moderate
            - low
        impact:
          type: string
          enum:
            - high
            - medium
            - low
        type:
          type: string
          enum:
            - standard
            - normal
            - emergency
            - model
            - site_reliability_ops
            - azure
        state:
          type: string
          enum:
            - new
            - assess
            - authorize
            - customer_approval
            - scheduled
            - implement
            - review
            - customer_review
            - rollback
            - closed
            - canceled
        groupId:
          type: string
          format: uuid
        assignedEngineerId:
          type: string
          format: uuid
        risk:
          type: string
          enum:
            - high
            - moderate
            - low
        requestedById:
          type: string
          format: uuid
        description:
          type: string
        justification:
          type: string
        implementationPlan:
          type: string
        riskImpactAnalysis:
          type: string
        backoutPlan:
          type: string
        testPlan:
          type: string
        plannedStartDate:
          type: string
          description: Planned start date and time (format YYYY-MM-DD HH:MM:SS).
          pattern: '^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}$'
        plannedEndDate:
          type: string
          description: Planned end date and time (format YYYY-MM-DD HH:MM:SS).
          pattern: '^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}$'
        comment:
          type: string
        workNote:
          type: string
    CreateChangeRequestResponse:
      type: object
      properties:
        message:
          type: string
        changeRequest:
          type: object
          properties:
            id:
              type: string
              format: uuid
            number:
              type: string
            createdOn:
              type: string
            createdBy:
              type: string
    PatchChangeRequestRequest:
      type: object
      minProperties: 1
      description: >
        Any combination of fields may be supplied; at least one is required.
        Omitted fields are left unchanged.
      properties:
        title:
          type: string
        description:
          type: string
        projectId:
          type: string
          format: uuid
        caseId:
          type: string
          format: uuid
        deploymentId:
          type: string
          format: uuid
        deployedProductId:
          type: string
          format: uuid
        assignedEngineerId:
          type: string
          format: uuid
        assignedTeamId:
          type: string
          format: uuid
        plannedStartOn:
          type: string
          description: Planned start date and time (format YYYY-MM-DD HH:MM:SS).
          pattern: '^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}$'
        plannedEndOn:
          type: string
          description: Planned end date and time (format YYYY-MM-DD HH:MM:SS).
          pattern: '^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}$'
        impact:
          type: string
          enum:
            - high
            - medium
            - low
        state:
          type: string
          enum:
            - new
            - assess
            - authorize
            - customer_approval
            - scheduled
            - implement
            - review
            - customer_review
            - rollback
            - closed
            - canceled
        type:
          type: string
          enum:
            - standard
            - normal
            - emergency
            - model
            - site_reliability_ops
            - azure
        justification:
          type: string
        impactDescription:
          type: string
        serviceOutage:
          type: string
        communicationPlan:
          type: string
        rollbackPlan:
          type: string
        testPlan:
          type: string
        isCustomerApproved:
          type: boolean
        isCustomerReviewed:
          type: boolean
    PatchChangeRequestResponse:
      type: object
      required:
        - message
        - changeRequest
      properties:
        message:
          type: string
        changeRequest:
          $ref: '#/components/schemas/ChangeRequest'
...
    SearchCommentsRequest:
      type: object
      required:
        - referenceId
        - referenceType
      properties:
        referenceId:
          type: string
          format: uuid
          description: UUID of the entity to fetch comments for.
        referenceType:
          type: string
          enum: [case, conversation, change_request, deployment]
          description: Type of the referenced entity.
        pagination:
          $ref: '#/components/schemas/Pagination'
    SearchCommentsResponse:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
          description: List of comments.
        total:
          type: integer
          description: Total number of comments matching the query.
        limit:
          type: integer
          description: Maximum number of comments returned in this page.
        offset:
          type: integer
          description: Number of comments skipped before this page.
        hasMore:
          type: boolean
          description: Whether additional pages are available.
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
...